Added divisions that signal the start of the next page/ Changed the title of the app/ Added a page id for my results/ Added a results button/ Added a fifth page for results/ Added a header on my last page/ Added a division to the last page/ Added identification that tells where to reveal results/ Added a function for my results/ Added functions to calculate the percentage of cold and allergy symptoms/ Added identification for my content of my results paragraph/ Added a way to calculate the probability of a cold or allergies
The divisions I added will take the survey from one page to the next. I also changed the title of the app to say "Cold or Allergy" instead of "Hello World". I added a button that will show the patient their results so they will be able to determine if they have a cold or allergy.